Add: source & tests files
Source files and tests files were added based on corresponding files in
perock library. _product.py & test_product.py copies of files in perock
library with minor changes.

This library prodius allows iterable to be used in cartesian product to be a
callable but also allowing iterable. Perock only expected iterables to only
contain callables or iterables not both. It was fixed in this commit and now
iterable can contain iterables and callables.

No matter how large iterable is, memory error will be avoided if iterable is
passed as callable e.g function. Intead of reading whole iterator into memory,
callable is called to get items of iterator again.

Using callable like function gives opportunity to change items of iterator that will
be used in each level. `itertools.product()` does not provide that particular behavious
as it expects iterable and its impossible to know if levels changed.

Limitation
----------
- Callbles or iterables are limited to 11 with repeat included.
- Lot of function calls which may lead to poor performance(not tested).

Issues
------
Problem is that current version of perock is licensed under GPL 3 but
this library(prodius) is licenced under MIT which is againt conditions
of GPL licences.

In the web its mentioned that GPL licence applies to users not authors and
the author still has copyright over the source code if they are only contributor.
No one contributed to perock besides me, that means I can do what I want with the
source.

That means I can reuse the source without restrictions of GPL which would force
this library to be released in that licence.